requirements.txt
opencv-python
pyzbar
qrcode
python-vlc